Application Analytics follow-up
