+++ apps/web/app/app.dub.co/(dashboard)/[slug]/(ee)/program/analytics/applications/application-referral-source-icon.tsx
@@ -1,0 +1,17 @@
+import { ReferrerIcon } from "@/ui/analytics/referrer-icon";
+import { Globe, Shop } from "@dub/ui/icons";
+
+export function ApplicationReferralSourceIcon({
+  referralSource,
+}: {
+  referralSource: string;
+}) {
+  switch (referralSource) {
+    case "marketplace":
+      return <Shop />;
+    case "direct":
+      return <Globe />;
+    default:
+      return <ReferrerIcon display={referralSource} />;
+  }
+}
