+++ apps/web/app/app.dub.co/(dashboard)/[slug]/(ee)/program/commissions/commission-menu-popover.tsx
@@ -1,13 +1,25 @@
 "use client";
 
+import useWorkspace from "@/lib/swr/use-workspace";
 import { useExportCommissionsModal } from "@/ui/modals/export-commissions-modal";
 import { ThreeDots } from "@/ui/shared/icons";
-import { Button, Download, IconMenu, Popover, Refresh2 } from "@dub/ui";
+import {
+  Button,
+  ChartLine,
+  Download,
+  IconMenu,
+  Popover,
+  Refresh2,
+  useRouterStuff,
+} from "@dub/ui";
+import Link from "next/link";
 import { useState } from "react";
 import { useCreateClawbackSheet } from "./create-clawback-sheet";
 
-export function CommissionPopoverButtons() {
+export function CommmissionsMenuPopover() {
   const [openPopover, setOpenPopover] = useState(false);
+  const { slug } = useWorkspace();
+  const { getQueryString } = useRouterStuff();
 
   const { createClawbackSheet, setIsOpen: setClawbackSheetOpen } =
     useCreateClawbackSheet({});
@@ -32,9 +44,31 @@ export function CommissionPopoverButtons() {
               >
                 <IconMenu
                   text="Create clawback"
-                  icon={<Refresh2 className="h-4 w-4" />}
+                  icon={<Refresh2 className="size-4" />}
                 />
               </button>
+              <Link
+                href={`/${slug}/program/analytics/commissions${getQueryString(
+                  undefined,
+                  {
+                    include: [
+                      "interval",
+                      "start",
+                      "end",
+                      "partnerId",
+                      "groupId",
+                      "type",
+                    ],
+                  },
+                )}`}
+              >
+                <button className="w-full rounded-md p-2 hover:bg-neutral-100 active:bg-neutral-200">
+                  <IconMenu
+                    text="View analytics"
+                    icon={<ChartLine className="size-4" />}
+                  />
+                </button>
+              </Link>
             </div>
 
             <div className="border-t border-neutral-200" />
@@ -52,7 +86,7 @@ export function CommissionPopoverButtons() {
               >
                 <IconMenu
                   text="Export as CSV"
-                  icon={<Download className="h-4 w-4" />}
+                  icon={<Download className="size-4" />}
                 />
               </button>
             </div>
