+++ apps/web/app/app.dub.co/(dashboard)/[slug]/(ee)/program/commissions/page.tsx
@@ -2,7 +2,7 @@
 
 import { PageContent } from "@/ui/layout/page-content";
 import { PageWidthWrapper } from "@/ui/layout/page-width-wrapper";
-import { CommissionPopoverButtons } from "./commission-popover-buttons";
+import { CommmissionsMenuPopover } from "./commission-menu-popover";
 import { CommissionsStats } from "./commissions-stats";
 import { CommissionsTable } from "./commissions-table";
 import { CreateCommissionButton } from "./create-commission-button";
@@ -19,7 +19,7 @@ export default function ProgramCommissions() {
       controls={
         <>
           <CreateCommissionButton />
-          <CommissionPopoverButtons />
+          <CommmissionsMenuPopover />
         </>
       }
     >
