+++ apps/web/app/app.dub.co/(dashboard)/[slug]/(ee)/program/partners/page.tsx
@@ -1,7 +1,7 @@
 import { PageContent } from "@/ui/layout/page-content";
 import { PageWidthWrapper } from "@/ui/layout/page-width-wrapper";
-import { ImportExportButtons } from "./import-export-buttons";
 import { InvitePartnerButton } from "./invite-partner-button";
+import { PartnersMenuPopover } from "./partners-menu-popover";
 import { PartnersTable } from "./partners-table";
 
 export default function ProgramPartners() {
@@ -16,7 +16,7 @@ export default function ProgramPartners() {
       controls={
         <>
           <InvitePartnerButton />
-          <ImportExportButtons />
+          <PartnersMenuPopover />
         </>
       }
     >
